requirements.txt
django==3.1.7
django-admin-sortable2==1.0
python-dotenv=0.15.0